Moved utils.py to dataframe
