+++ modin/experimental/core/execution/native/implementations/hdk_on_native/dataframe/utils.py
@@ -10,6 +10,10 @@
 # the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific language
 # governing permissions and limitations under the License.
+
+"""Utilities for internal use by te hdk_on_native module."""
+
+
 import pandas
 import pyarrow
